pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wiseflow"
version = "4.30"
description = "WiseFlow - Use LLMs to track and extract websites, RSS feeds, and social media"
authors = [
    {name = "TeamWiseflow", email = "info@wiseflow.pro"},
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12,<3.13"
keywords = ["web-scraping", "llm", "crawler", "ai", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Logging system
    "loguru>=0.7.2",
    # Async HTTP clients
    "httpx==0.28.1",
    "requests>=2.32.3",
    # Database and storage
    "aiosqlite>=0.20.0",
    # Data validation and processing
    "pydantic>=2.10.3",
    "pandas>=2.2.3",
    "parsel==1.6.0",
    "pyhumps==3.8.0",
    # LLM API
    "openai>=1.58.1",
    # Browser automation
    "patchright==1.52.5",
    "nodriver==0.47",
    "pdf2image>=1.17.0",
    # HTML/XML processing
    "beautifulsoup4~=4.12",
    "lxml~=5.3",
    "cssselect==1.1.0", # compatiled with mediacrawler-pro
    # Text processing
    "regex>=2024.11.6",
    "snowballstemmer~=2.2",
    # RSS and data parsing
    "feedparser>=6.0.11",
    # Image processing
    "pillow>=10.4.0",
    # Environment and configuration
    "python-dotenv~=1.0",
    "packaging>=24.2",
    # File handling
    "aiofiles==24.1.0",
    # Hashing and encryption
    "xxhash>=3.5.0",
    "pyOpenSSL>=24.3.0",
    "cryptography",
    # System and performance
    "psutil>=6.1.1",
    "lark>=1.1.7",
    # SignSrv requirements
    "PyExecJS==1.5.1",
    # Character encoding detection
    "faust-cchardet>=2.1.19",
    "chardet>=5.2.0",
    # Retry mechanism
    "tenacity==9.0.0", # Compatible with SignSrv
    "certifi>=2024.12.14",
    # Office file processing
    "openpyxl",
    # Task scheduling
    "schedule>=1.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Authentication
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "black>=24.10.0",
]

[project.scripts]
wiseflow = "core.run_task:main"
wiseflow-backend = "core.backend.main:main"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[project.urls]
Homepage = "https://github.com/teamwiseflow/wiseflow"
Repository = "https://github.com/teamwiseflow/wiseflow"
Issues = "https://github.com/teamwiseflow/wiseflow/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "black>=24.10.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PIE", "SIM"]
ignore = ["E501", "B008", "B904"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 120
target-version = ['py312']

[index]
url = "https://mirrors.aliyun.com/pypi/simple/"
